1.报销问题修复及样式修改  2.报名表样式修改

+++ src/com/community/bean/Audit.java
@@ -16,7 +16,7 @@ public class Audit {
 	private String department;
 	private Date time;
 	private String reason;
-	private Integer total;
+	private Float total;
 	private Integer state;
 	private String procInstId;
 	private String audit_suggest;
@@ -58,10 +58,10 @@ public class Audit {
 	public void setReason(String reason) {
 		this.reason = reason;
 	}
-	public Integer getTotal() {
+	public Float getTotal() {
 		return total;
 	}
-	public void setTotal(Integer total) {
+	public void setTotal(Float total) {
 		this.total = total;
 	}
 	public Integer getState() {
